requirements.txt
# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
tenacity>=8.0.0
docker>=7.0.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.0.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# Configuration Management (NEW - for vendor configs)
PyYAML>=6.0.0

# Flask Application
Flask>=3.0.0
Flask-CORS>=4.0.0
sentry-sdk>=2.0.0

# Database
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
psycopg[binary]>=3.2.0
cryptography>=42.0.0
celery>=5.3.0
flower>=2.0.0

# Authentication
flask-login==0.6.3
flask-session[redis]==0.8.0
flask-bcrypt==1.0.1
redis==5.0.3
itsdangerous==2.1.2
flask-mail==0.10.0
stripe==11.1.0

# API Infrastructure
flask-openapi3[swagger]==4.3.1
flask-limiter==4.1.1
flask-compress==1.15.0

# Excel Support
openpyxl>=3.1.0

# Optional: Advanced Features
# python-slugify>=8.0.0  # SEO-friendly slugs
# deep-translator>=1.11.0  # Multilingual support
Pillow>=10.0.0  # Image processing (SVSE audit)
ImageHash>=4.3.1  # Perceptual hashing
open-clip-torch>=2.24.0  # CLIP embeddings
torch>=2.1.0  # CLIP runtime
numpy>=1.26.0  # Embedding arrays
scikit-learn>=1.3.0  # Clustering
sentence-transformers>=2.0.0  # Product embeddings for semantic search

# CLI Framework
typer>=0.21.0
rich>=14.0.0

# Development & Testing
psutil>=5.9.0
pytest==8.3.5
pytest-cov==6.1.0
pytest-asyncio==0.25.3
requests-mock==1.12.1

# Graph Database and Temporal Knowledge (Phase 13.2)
graphiti-core==0.26.0
neo4j==5.26.0
mcp==1.12.4
